+++ gulpfile.js
@@ -61,7 +61,7 @@ gulp.task(`sprite`, () => {
 });
 
 gulp.task(`scripts`, () => {
-  return gulp.src(`js/application.js`)
+  return gulp.src(`js/main.js`)
     .pipe(plumber())
     .pipe(sourcemaps.init())
     .pipe(rollup({}, `iife`))
