+++ js/components/abstract-view.js
