+++ js/main.js
@@ -1,0 +1,3 @@
+import Application from './presenters/application.js';
+
+Application.showIntro();
